src/Brain/Confidence/Discrete.java
package Brain.Confidence;


public enum Discrete {
    ZERO, ONE, TWO, THREE, FOUR, FIVE, SIX, SEVEN, EIGHT, NINE, TEN;

    public static Discrete discrete(double value) {
        int x10 = (int) (value * 10);
        switch (x10) {
            case 0:
                return ZERO;
            case 1:
                return ONE;
            case 2:
                return TWO;
            case 3:
                return THREE;
            case 4:
                return FOUR;
            case 5:
                return FIVE;
            case 6 :
                return SIX;
            case 7:
                return SEVEN;
            case 8:
                return EIGHT;
            case 9:
                return NINE;
            case 10:
                return TEN;
            default:
                throw new RuntimeException("The code should not arrive here");
        }
    }
}